Correction: corrigido o nome incorreto da class library.

+++ Controller/GerenciadorDeSenhas.Controller/TransacaoController.cs
@@ -10,7 +10,7 @@ namespace GerenciadorDeSenhas.Controller
 {
     public class TransacaoController
     {
-        public bool InserirTransacao(string nomeTransacao, string tipoTransacao, DateTime dataTransacao, string descTransacao, double valorTransacao)
+        public bool InserirTransacao(string nomeTransacao, string tipoTransacao, DateTime dataTransacao, string descTransacao, decimal valorTransacao)
         {
             var repositorio = new TrasacaoRepository();
             var resposta = repositorio.Inserir(nomeTransacao, tipoTransacao, dataTransacao, descTransacao, valorTransacao);
